feat: добавлена лента событий
