+++ src/main/java/ru/yandex/practicum/filmorate/dao/user/UserDAO.java
@@ -11,6 +11,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Repository;
+import ru.yandex.practicum.filmorate.dao.feed.FeedDAO;
 import ru.yandex.practicum.filmorate.exception.friend.FriendAlreadyAddedException;
 import ru.yandex.practicum.filmorate.exception.friend.FriendNotFoundException;
 import ru.yandex.practicum.filmorate.exception.user.UserNotFoundException;
@@ -26,6 +27,7 @@ import java.util.Collection;
 public class UserDAO implements UserStorage {
 
     private final JdbcTemplate template;
+    private final FeedDAO feed;
 
     @Override
     public User getUser(int id) {
@@ -40,7 +42,10 @@ public class UserDAO implements UserStorage {
         SqlRowSet rowSet = template.queryForRowSet(
                 "SELECT \"to_user_id\" FROM \"friendships_sent\" WHERE \"from_user_id\" = ?", id);
         while (rowSet.next()) {
-            user.addFriend(rowSet.getInt("to_user_id"));
+            if (user != null) {
+                user.addFriend(rowSet.getInt("to_user_id"));
+
+            } else throw new RuntimeException("Ошибка при получении пользователя.");
         }
 
         return user;
@@ -81,6 +86,7 @@ public class UserDAO implements UserStorage {
                 targetUserId, friendId);
 
         if (rowSet.next()) {
+            feed.addFriendEvent(targetUserId, friendId);
             throw new FriendAlreadyAddedException(
                     String.format("Ошибка при добавлении друга для пользователя с id=%d: " +
                             "друг с id=%d уже добавлен.", targetUserId, friendId)
@@ -90,6 +96,7 @@ public class UserDAO implements UserStorage {
         try {
             template.update("INSERT INTO \"friendships_sent\" (\"from_user_id\", \"to_user_id\")" +
                     "VALUES (?, ?)", targetUserId, friendId);
+            feed.addFriendEvent(targetUserId, friendId);
         } catch (DataIntegrityViolationException e) {
             throw new UserNotFoundException(String.format("Ошибка при добавлении друга с id=%d" +
                     " для пользователя с id=%d: один или оба пользователя не найдены.", friendId, targetUserId));
@@ -135,6 +142,7 @@ public class UserDAO implements UserStorage {
         }
         template.update("DELETE FROM \"friendships_sent\" WHERE \"from_user_id\" = ? AND \"to_user_id\" = ?",
                 targetUserId, friendId);
+        feed.removeFriendEvent(targetUserId, friendId);
     }
 
     @Override
