+++ src/main/java/ru/yandex/practicum/filmorate/service/user/UserService.java
@@ -1,5 +1,6 @@
 package ru.yandex.practicum.filmorate.service.user;
 
+import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.Collection;
@@ -13,6 +14,8 @@ public interface UserService {
     Collection<User> getFriendsList(int targetUserId);
 
     Collection<User> getCommonFriends(int targetUserId, int otherUserId);
+
+    Collection<Feed> getUserFeed(int userId);
 
     int addUser(User user);
