+++ src/main/java/ru/yandex/practicum/filmorate/service/user/impl/InMemoryUserService.java
@@ -2,11 +2,14 @@ package ru.yandex.practicum.filmorate.service.user.impl;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.dao.feed.FeedDAO;
+import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.user.UserService;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -14,9 +17,12 @@ import java.util.stream.Collectors;
 public class InMemoryUserService implements UserService {
 
     protected final UserStorage storage;
+    protected final FeedDAO feed;
 
-    public InMemoryUserService(@Qualifier("InMemoryUserStorage") UserStorage storage) {
+    public InMemoryUserService(@Qualifier("InMemoryUserStorage") UserStorage storage,
+                               @Qualifier("FeedDAO") FeedDAO feed) {
         this.storage = storage;
+        this.feed = feed;
     }
 
     @Override
@@ -49,6 +55,13 @@ public class InMemoryUserService implements UserService {
                 .map(storage::getUser)
                 .collect(Collectors.toList())
         );
+    }
+
+    @Override
+    public Collection<Feed> getUserFeed(int userId) {
+        return feed.getUserFeed(userId).stream()
+                .sorted(Comparator.comparing(Feed::getEventId))
+                .collect(Collectors.toList());
     }
 
     @Override
