+++ src/main/java/ru/yandex/practicum/filmorate/service/user/impl/UserDBService.java
@@ -3,14 +3,16 @@ package ru.yandex.practicum.filmorate.service.user.impl;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.dao.feed.FeedDAO;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 @Service("UserDBService")
 @Primary
 public class UserDBService extends InMemoryUserService {
 
-    public UserDBService(@Qualifier("UserDAO") UserStorage storage) {
-        super(storage);
+    public UserDBService(@Qualifier("UserDAO") UserStorage storage,
+                         @Qualifier("FeedDAO") FeedDAO feed) {
+        super(storage, feed);
     }
 
     @Override
