+++ src/main/resources/schema.sql
@@ -103,3 +103,25 @@ CREATE TABLE IF NOT EXISTS "review_like_dislike"
     "is_like"   BOOLEAN,
     PRIMARY KEY ("review_id", "user_id", "is_like")
 );
+
+CREATE TYPE IF NOT EXISTS "event_type" AS ENUM (
+    'LIKE',
+    'REVIEW',
+    'FRIEND'
+    );
+
+CREATE TYPE IF NOT EXISTS "operation_type" AS ENUM (
+    'REMOVE',
+    'ADD',
+    'UPDATE'
+    );
+
+CREATE TABLE IF NOT EXISTS "feed"
+(
+    "timestamp" TIMESTAMP NOT NULL,
+    "user_id"   INT REFERENCES "users" ("user_id") ON DELETE CASCADE,
+    "eventType" "event_type" NOT NULL,
+    "operation" "operation_type" NOT NULL,
+    "event_id" INT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+    "entity_id" INT NOT NULL
+);
