+++ src/test/resources/schematest.sql
@@ -42,7 +42,6 @@ CREATE TABLE IF NOT EXISTS "films"
     "duration"     INT          NOT NULL,
     "rate"         INT          NOT NULL,
     "mpa_id"       INT REFERENCES "mpa_rating" ("mpa_id"),
-    "director_id"  INT          NULL,
     CONSTRAINT IF NOT EXISTS "not_blank" CHECK (LENGTH("film_name") > 0),
     CONSTRAINT IF NOT EXISTS "earliest_date" CHECK ("release_date" >= CAST('1895-12-28' AS DATE)),
     CONSTRAINT IF NOT EXISTS "positive" CHECK ("duration" > 0)
@@ -103,4 +102,26 @@ CREATE TABLE IF NOT EXISTS "review_like_dislike"
     "user_id"   INT REFERENCES "users" ("user_id") ON DELETE CASCADE,
     "is_like"   BOOLEAN,
     PRIMARY KEY ("review_id", "user_id", "is_like")
+);
+
+CREATE TYPE IF NOT EXISTS "event_type" AS ENUM (
+    'LIKE',
+    'REVIEW',
+    'FRIEND'
+    );
+
+CREATE TYPE IF NOT EXISTS "operation_type" AS ENUM (
+    'REMOVE',
+    'ADD',
+    'UPDATE'
+    );
+
+CREATE TABLE IF NOT EXISTS "feed"
+(
+    "timestamp" TIMESTAMP NOT NULL,
+    "user_id"   INT REFERENCES "users" ("user_id") ON DELETE CASCADE,
+    "eventType" "event_type" NOT NULL,
+    "operation" "operation_type" NOT NULL,
+    "event_id" INT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+    "entity_id" INT NOT NULL
 );
